pyproject.toml
[project]
name = "turbo-review"
version = "0.1.0"
description = "AI-powered code review system with GraphRAG"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "openai>=1.3.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "tree-sitter>=0.20.0",
    "tree-sitter-python>=0.20.0",
    "tree-sitter-javascript>=0.20.0",
    "tree-sitter-typescript>=0.20.0",
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "aiohttp>=3.8.0",
    "asyncio>=3.4.3",
    "unidiff>=0.7.5",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation>=0.41b0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-instrumentation-aiohttp-client>=0.41b0",
    "opentelemetry-instrumentation-requests>=0.41b0",
    "opentelemetry-instrumentation-httpx>=0.41b0",
    "networkx>=3.0",
    "python-louvain>=0.16",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cli", "core", "processing", "tests", "utils", "integrations", "inference", "monitoring", "services"]

[project.scripts]
turbo-review = "main:main"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:builtin type SwigPyPacked",
    "ignore::DeprecationWarning:builtin type SwigPyObject",
    "ignore::DeprecationWarning:builtin type swigvarlink",
]
